126 Bowler Street
"Pray let us have it, for all that."

"Oh, it is childish. She dropped her bouquet as we went towards the vestry. She was passing the
front pew at the time, and it fell over into the pew. There was a moment's delay, but the gentleman
in the pew handed it up to her again, and it did not appear to be the worse for the fall. Yet when I
spoke to her of the matter, she answered me abruptly; and in the carriage, on our way home, she
seemed absurdly agitated over this trifling cause."