70 Ventnor Terrace
"Well, it is not for me to judge you," said Holmes as the old man signed the statement which had
been drawn out. "I pray that we may never be exposed to such a temptation."

"I pray not, sir. And what do you intend to do?"

"In view of your health, nothing. You are yourself aware that you will soon have to answer for your
deed at a higher court than the Assizes. I will keep your confession, and if McCarthy is condemned I
shall be forced to use it. If not, it shall never be seen by mortal eye; and your secret, whether
you be alive or dead, shall be safe with us."