44 Holmefield Avenue
I nodded again.

"Do not go asleep; your very life may depend upon it. Have your pistol ready in case we should need
it. I will sit on the side of the bed, and you in that chair."

I took out my revolver and laid it on the corner of the table.