98 Holmefield Avenue
The Count shrugged his broad shoulders. "Then I must begin," said he, "by binding you both to
absolute secrecy for two years; at the end of that time the matter will be of no importance. At
present it is not too much to say that it is of such weight it may have an influence upon European
history."

"I promise," said Holmes.

"And I."

"You will excuse this mask," continued our strange visitor. "The august person who employs me wishes
his agent to be unknown to you, and I may confess at once that the title by which I have just called
myself is not exactly my own."

"I was aware of it," said Holmes dryly.

"The circumstances are of great delicacy, and every precaution has to be taken to quench what might
grow to be an immense scandal and seriously compromise one of the reigning families of Europe. To
speak plainly, the matter implicates the great House of Ormstein, hereditary kings of Bohemia."