34 Larch Walk
"Oh, to the other side of the West End. It may be some time before I get back. Don't wait up for me
in case I should be late."

"How are you getting on?"

"Oh, so so. Nothing to complain of. I have been out to Streatham since I saw you last, but I did not
call at the house. It is a very sweet little problem, and I would not have missed it for a good
deal. However, I must not sit gossiping here, but must get these disreputable clothes off and return
to my highly respectable self."

I could see by his manner that he had stronger reasons for satisfaction than his words alone would
imply. His eyes twinkled, and there was even a touch of colour upon his sallow cheeks. He hastened
upstairs, and a few minutes later I heard the slam of the hall door, which told me that he was off
once more upon his congenial hunt.

I waited until midnight, but there was no sign of his return, so I retired to my room. It was no
uncommon thing for him to be away for days and nights on end when he was hot upon a scent, so that
his lateness caused me no surprise. I do not know at what hour he came in, but when I came down to
breakfast in the morning there he was with a cup of coffee in one hand and the paper in the other,
as fresh and trim as possible.

"You will excuse my beginning without you, Watson," said he, "but you remember that our client has
rather an early appointment this morning."