161 Wantage Close
"Oh, no, no! I want the doctor's advice and help, too. It's about Isa. He has not been home for two
days. I am so frightened about him!"

It was not the first time that she had spoken to us of her husband's trouble, to me as a doctor, to
my wife as an old friend and school companion. We soothed and comforted her by such words as we
could find. Did she know where her husband was? Was it possible that we could bring him back to her?