123 Wantage Close
"I cannot admire his taste," I remarked, "if it is indeed a fact that he was averse to a marriage
with so charming a young lady as this Miss Turner."

"Ah, thereby hangs a rather painful tale. This fellow is madly, insanely, in love with her, but some
two years ago, when he was only a lad, and before he really knew her, for she had been away five
years at a boarding-school, what does the idiot do but get into the clutches of a barmaid in Bristol
and marry her at a registry office? No one knows a word of the matter, but you can imagine how
maddening it must be to him to be upbraided for not doing what he would give his very eyes to do,
but what he knows to be absolutely impossible. It was sheer frenzy of this sort which made him throw
his hands up into the air when his father, at their last interview, was goading him on to propose to
Miss Turner. On the other hand, he had no means of supporting himself, and his father, who was by
all accounts a very hard man, would have thrown him over utterly had he known the truth. It was with
his barmaid wife that he had spent the last three days in Bristol, and his father did not know where
he was. Mark that point. It is of importance. Good has come out of evil, however, for the barmaid,
finding from the papers that he is in serious trouble and likely to be hanged, has thrown him over
utterly and has written to him to say that she has a husband already in the Bermuda Dockyard, so
that there is really no tie between them. I think that that bit of news has consoled young McCarthy
for all that he has suffered."

"But if he is innocent, who has done it?"

"Ah! who? I would call your attention very particularly to two points. One is that the murdered man
had an appointment with someone at the pool, and that the someone could not have been his son, for
his son was away, and he did not know when he would return. The second is that the murdered man was
heard to cry 'Cooee!' before he knew that his son had returned. Those are the crucial points upon
which the case depends. And now let us talk about George Meredith, if you please, and we shall leave
all minor matters until to-morrow."

There was no rain, as Holmes had foretold, and the morning broke bright and cloudless. At nine
o'clock Lestrade called for us with the carriage, and we set off for Hatherley Farm and the Boscombe
Pool.

"There is serious news this morning," Lestrade observed. "It is said that Mr. Turner, of the Hall,
is so ill that his life is despaired of."