101 Moor End
"Yes, but I do not think that it is such a very unusual thing to have a small opening between two
rooms. It was so small that a rat could hardly pass through."

"I knew that we should find a ventilator before ever we came to Stoke Moran."

"My dear Holmes!"

"Oh, yes, I did. You remember in her statement she said that her sister could smell Dr. Roylott's
cigar. Now, of course that suggested at once that there must be a communication between the two
rooms. It could only be a small one, or it would have been remarked upon at the coroner's inquiry. I
deduced a ventilator."