19 Balcombe Close
"Their reasons for their conduct. But you shall have it all just as it occurred. When I came down,
Mr. Rucastle met me here and drove me in his dog-cart to the Copper Beeches. It is, as he said,
beautifully situated, but it is not beautiful in itself, for it is a large square block of a house,
whitewashed, but all stained and streaked with damp and bad weather. There are grounds round it,
woods on three sides, and on the fourth a field which slopes down to the Southampton highroad, which
curves past about a hundred yards from the front door. This ground in front belongs to the house,
but the woods all round are part of Lord Southerton's preserves. A clump of copper beeches
immediately in front of the hall door has given its name to the place.

"I was driven over by my employer, who was as amiable as ever, and was introduced by him that
evening to his wife and the child. There was no truth, Mr. Holmes, in the conjecture which seemed to
us to be probable in your rooms at Baker Street. Mrs. Rucastle is not mad. I found her to be a
silent, pale-faced woman, much younger than her husband, not more than thirty, I should think, while
he can hardly be less than forty-five. From their conversation I have gathered that they have been
married about seven years, that he was a widower, and that his only child by the first wife was the
daughter who has gone to Philadelphia. Mr. Rucastle told me in private that the reason why she had
left them was that she had an unreasoning aversion to her stepmother. As the daughter could not have
been less than twenty, I can quite imagine that her position must have been uncomfortable with her
father's young wife.

"Mrs. Rucastle seemed to me to be colourless in mind as well as in feature. She impressed me neither
favourably nor the reverse. She was a nonentity. It was easy to see that she was passionately
devoted both to her husband and to her little son. Her light grey eyes wandered continually from one
to the other, noting every little want and forestalling it if possible. He was kind to her also in
his bluff, boisterous fashion, and on the whole they seemed to be a happy couple. And yet she had
some secret sorrow, this woman. She would often be lost in deep thought, with the saddest look upon
her face. More than once I have surprised her in tears. I have thought sometimes that it was the
disposition of her child which weighed upon her mind, for I have never met so utterly spoiled and so
ill-natured a little creature. He is small for his age, with a head which is quite
disproportionately large. His whole life appears to be spent in an alternation between savage fits
of passion and gloomy intervals of sulking. Giving pain to any creature weaker than himself seems to
be his one idea of amusement, and he shows quite remarkable talent in planning the capture of mice,
little birds, and insects. But I would rather not talk about the creature, Mr. Holmes, and, indeed,
he has little to do with my story."

"I am glad of all details," remarked my friend, "whether they seem to you to be relevant or not."

"I shall try not to miss anything of importance. The one unpleasant thing about the house, which
struck me at once, was the appearance and conduct of the servants. There are only two, a man and his
wife. Toller, for that is his name, is a rough, uncouth man, with grizzled hair and whiskers, and a
perpetual smell of drink. Twice since I have been with them he has been quite drunk, and yet Mr.
Rucastle seemed to take no notice of it. His wife is a very tall and strong woman with a sour face,
as silent as Mrs. Rucastle and much less amiable. They are a most unpleasant couple, but fortunately
I spend most of my time in the nursery and my own room, which are next to each other in one corner
of the building.

"For two days after my arrival at the Copper Beeches my life was very quiet; on the third, Mrs.
Rucastle came down just after breakfast and whispered something to her husband.