87 Harvey Street
"I did as I was told, and at the same instant Mrs. Rucastle drew down the blind. That was a week
ago, and from that time I have not sat again in the window, nor have I worn the blue dress, nor seen
the man in the road."

"Pray continue," said Holmes. "Your narrative promises to be a most interesting one."