65 Beaconside
"'At least,' said he, 'you will not have me arrested at once. It would be to your advantage as well
as mine if I might leave the house for five minutes.'

"'That you may get away, or perhaps that you may conceal what you have stolen,' said I. And then,
realising the dreadful position in which I was placed, I implored him to remember that not only my
honour but that of one who was far greater than I was at stake; and that he threatened to raise a
scandal which would convulse the nation. He might avert it all if he would but tell me what he had
done with the three missing stones.