13 Down Close
By eleven o'clock the next day we were well upon our way to the old English capital. Holmes had been
buried in the morning papers all the way down, but after we had passed the Hampshire border he threw
them down and began to admire the scenery. It was an ideal spring day, a light blue sky, flecked
with little fleecy white clouds drifting across from west to east. The sun was shining very
brightly, and yet there was an exhilarating nip in the air, which set an edge to a man's energy. All
over the countryside, away to the rolling hills around Aldershot, the little red and grey roofs of
the farm-steadings peeped out from amid the light green of the new foliage.

"Are they not fresh and beautiful?" I cried with all the enthusiasm of a man fresh from the fogs of
Baker Street.

But Holmes shook his head gravely.

"Do you know, Watson," said he, "that it is one of the curses of a mind with a turn like mine that I
must look at everything with reference to my own special subject. You look at these scattered
houses, and you are impressed by their beauty. I look at them, and the only thought which comes to
me is a feeling of their isolation and of the impunity with which crime may be committed there."

"Good heavens!" I cried. "Who would associate crime with these dear old homesteads?"

"They always fill me with a certain horror. It is my belief, Watson, founded upon my experience,
that the lowest and vilest alleys in London do not present a more dreadful record of sin than does
the smiling and beautiful countryside."