111 Moor End
"And they were posted to-day at Gravesend. Well, Mrs. St. Clair, the clouds lighten, though I should
not venture to say that the danger is over."

"But he must be alive, Mr. Holmes."

"Unless this is a clever forgery to put us on the wrong scent. The ring, after all, proves nothing.
It may have been taken from him."

"No, no; it is, it is his very own writing!"