60 Tamworth Drive
"And yet," said I, smiling, "I cannot quite hold myself absolved from the charge of sensationalism
which has been urged against my records."

"You have erred, perhaps," he observed, taking up a glowing cinder with the tongs and lighting with
it the long cherry-wood pipe which was wont to replace his clay when he was in a disputatious rather
than a meditative mood--"you have erred perhaps in attempting to put colour and life into each of
your statements instead of confining yourself to the task of placing upon record that severe
reasoning from cause to effect which is really the only notable feature about the thing."

"It seems to me that I have done you full justice in the matter," I remarked with some coldness, for
I was repelled by the egotism which I had more than once observed to be a strong factor in my
friend's singular character.

"No, it is not selfishness or conceit," said he, answering, as was his wont, my thoughts rather than
my words. "If I claim full justice for my art, it is because it is an impersonal thing--a thing
beyond myself. Crime is common. Logic is rare. Therefore it is upon the logic rather than upon the
crime that you should dwell. You have degraded what should have been a course of lectures into a
series of tales."