6 Down Close
"They often vanish before the ceremony, and occasionally during the honeymoon; but I cannot call to
mind anything quite so prompt as this. Pray let me have the details."

"I warn you that they are very incomplete."

"Perhaps we may make them less so."

"Such as they are, they are set forth in a single article of a morning paper of yesterday, which I
will read to you. It is headed, 'Singular Occurrence at a Fashionable Wedding':