2 Glan Road
"Well, it is conjectured to be so. I shall take nothing for granted until I have the opportunity of
looking personally into it. I will explain the state of things to you, as far as I have been able to
understand it, in a very few words.

"Boscombe Valley is a country district not very far from Ross, in Herefordshire. The largest landed
proprietor in that part is a Mr. John Turner, who made his money in Australia and returned some
years ago to the old country. One of the farms which he held, that of Hatherley, was let to Mr.
Charles McCarthy, who was also an ex-Australian. The men had known each other in the colonies, so
that it was not unnatural that when they came to settle down they should do so as near each other as
possible. Turner was apparently the richer man, so McCarthy became his tenant but still remained, it
seems, upon terms of perfect equality, as they were frequently together. McCarthy had one son, a lad
of eighteen, and Turner had an only daughter of the same age, but neither of them had wives living.
They appear to have avoided the society of the neighbouring English families and to have led retired
lives, though both the McCarthys were fond of sport and were frequently seen at the race-meetings of
the neighbourhood. McCarthy kept two servants--a man and a girl. Turner had a considerable
household, some half-dozen at the least. That is as much as I have been able to gather about the
families. Now for the facts.