80 Beaconside
"Hum! So much for the police-court," said Holmes thoughtfully, tossing aside the paper. "The
question for us now to solve is the sequence of events leading from a rifled jewel-case at one end
to the crop of a goose in Tottenham Court Road at the other. You see, Watson, our little deductions
have suddenly assumed a much more important and less innocent aspect. Here is the stone; the stone
came from the goose, and the goose came from Mr. Henry Baker, the gentleman with the bad hat and all
the other characteristics with which I have bored you. So now we must set ourselves very seriously
to finding this gentleman and ascertaining what part he has played in this little mystery. To do
this, we must try the simplest means first, and these lie undoubtedly in an advertisement in all the
evening papers. If this fail, I shall have recourse to other methods."

"What will you say?"

"Give me a pencil and that slip of paper. Now, then: 'Found at the corner of Goodge Street, a goose
and a black felt hat. Mr. Henry Baker can have the same by applying at 6.30 this evening at 221b,
Baker Street.' That is clear and concise."

"Very. But will he see it?"

"Well, he is sure to keep an eye on the papers, since, to a poor man, the loss was a heavy one. He
was clearly so scared by his mischance in breaking the window and by the approach of Peterson that
he thought of nothing but flight, but since then he must have bitterly regretted the impulse which
caused him to drop his bird. Then, again, the introduction of his name will cause him to see it, for
everyone who knows him will direct his attention to it. Here you are, Peterson, run down to the
advertising agency and have this put in the evening papers."