86 Wilson Gardens
"Entirely."

"Do not think of revenge, or anything of the sort, at present. I think that we may gain that by
means of the law; but we have our web to weave, while theirs is already woven. The first
consideration is to remove the pressing danger which threatens you. The second is to clear up the
mystery and to punish the guilty parties."