Crime scene report #791049
"By the select prices. Eight shillings for a bed and eightpence for a glass of sherry pointed to one
of the most expensive hotels. There are not many in London which charge at that rate. In the second
one which I visited in Northumberland Avenue, I learned by an inspection of the book that Francis H.
Moulton, an American gentleman, had left only the day before, and on looking over the entries
against him, I came upon the very items which I had seen in the duplicate bill. His letters were to
be forwarded to 226 Gordon Square; so thither I travelled, and being fortunate enough to find the
loving couple at home, I ventured to give them some paternal advice and to point out to them that it
would be better in every way that they should make their position a little clearer both to the
general public and to Lord St. Simon in particular. I invited them to meet him here, and, as you
see, I made him keep the appointment."

"But with no very good result," I remarked. "His conduct was certainly not very gracious."

"Ah, Watson," said Holmes, smiling, "perhaps you would not be very gracious either, if, after all
the trouble of wooing and wedding, you found yourself deprived in an instant of wife and of fortune.
I think that we may judge Lord St. Simon very mercifully and thank our stars that we are never
likely to find ourselves in the same position. Draw your chair up and hand me my violin, for the
only problem we have still to solve is how to while away these bleak autumnal evenings."

"Holmes," said I as I stood one morning in our bow-window looking down the street, "here is a madman
coming along. It seems rather sad that his relatives should allow him to come out alone."

My friend rose lazily from his armchair and stood with his hands in the pockets of his dressing-
gown, looking over my shoulder. It was a bright, crisp February morning, and the snow of the day
before still lay deep upon the ground, shimmering brightly in the wintry sun. Down the centre of
Baker Street it had been ploughed into a brown crumbly band by the traffic, but at either side and
on the heaped-up edges of the foot-paths it still lay as white as when it fell. The grey pavement
had been cleaned and scraped, but was still dangerously slippery, so that there were fewer
passengers than usual. Indeed, from the direction of the Metropolitan Station no one was coming save
the single gentleman whose eccentric conduct had drawn my attention.