Crime scene report #308842
"Then I shall call upon you in a day, or in two days, with news as to the box and the papers. I
shall take your advice in every particular." He shook hands with us and took his leave. Outside the
wind still screamed and the rain splashed and pattered against the windows. This strange, wild story
seemed to have come to us from amid the mad elements--blown in upon us like a sheet of sea-weed in a
gale--and now to have been reabsorbed by them once more.

Sherlock Holmes sat for some time in silence, with his head sunk forward and his eyes bent upon the
red glow of the fire. Then he lit his pipe, and leaning back in his chair he watched the blue smoke-
rings as they chased each other up to the ceiling.