Crime scene report #678579
"And it did, though they hardly found upon the mud-bank what they had feared to find. It was Neville
St. Clair's coat, and not Neville St. Clair, which lay uncovered as the tide receded. And what do
you think they found in the pockets?"

"I cannot imagine."

"No, I don't think you would guess. Every pocket stuffed with pennies and half-pennies--421 pennies
and 270 half-pennies. It was no wonder that it had not been swept away by the tide. But a human body
is a different matter. There is a fierce eddy between the wharf and the house. It seemed likely
enough that the weighted coat had remained when the stripped body had been sucked away into the
river."

"But I understand that all the other clothes were found in the room. Would the body be dressed in a
coat alone?"