Crime scene report #605945
"I can't imagine. I suppose that you have been watching the habits, and perhaps the house, of Miss
Irene Adler."

"Quite so; but the sequel was rather unusual. I will tell you, however. I left the house a little
after eight o'clock this morning in the character of a groom out of work. There is a wonderful
sympathy and freemasonry among horsey men. Be one of them, and you will know all that there is to
know. I soon found Briony Lodge. It is a bijou villa, with a garden at the back, but built out in
front right up to the road, two stories. Chubb lock to the door. Large sitting-room on the right
side, well furnished, with long windows almost to the floor, and those preposterous English window
fasteners which a child could open. Behind there was nothing remarkable, save that the passage
window could be reached from the top of the coach-house. I walked round it and examined it closely
from every point of view, but without noting anything else of interest.

"I then lounged down the street and found, as I expected, that there was a mews in a lane which runs
down by one wall of the garden. I lent the ostlers a hand in rubbing down their horses, and received
in exchange twopence, a glass of half and half, two fills of shag tobacco, and as much information
as I could desire about Miss Adler, to say nothing of half a dozen other people in the neighbourhood
in whom I was not in the least interested, but whose biographies I was compelled to listen to."

"And what of Irene Adler?" I asked.

"Oh, she has turned all the men's heads down in that part. She is the daintiest thing under a bonnet
on this planet. So say the Serpentine-mews, to a man. She lives quietly, sings at concerts, drives
out at five every day, and returns at seven sharp for dinner. Seldom goes out at other times, except
when she sings. Has only one male visitor, but a good deal of him. He is dark, handsome, and
dashing, never calls less than once a day, and often twice. He is a Mr. Godfrey Norton, of the Inner
Temple. See the advantages of a cabman as a confidant. They had driven him home a dozen times from
Serpentine-mews, and knew all about him. When I had listened to all they had to tell, I began to
walk up and down near Briony Lodge once more, and to think over my plan of campaign.