Crime scene report #139024
"Well, down they came to the west country, there was no shaking them off, and there they have lived
rent free on my best land ever since. There was no rest for me, no peace, no forgetfulness; turn
where I would, there was his cunning, grinning face at my elbow. It grew worse as Alice grew up, for
he soon saw I was more afraid of her knowing my past than of the police. Whatever he wanted he must
have, and whatever it was I gave him without question, land, money, houses, until at last he asked a
thing which I could not give. He asked for Alice.

"His son, you see, had grown up, and so had my girl, and as I was known to be in weak health, it
seemed a fine stroke to him that his lad should step into the whole property. But there I was firm.
I would not have his cursed stock mixed with mine; not that I had any dislike to the lad, but his
blood was in him, and that was enough. I stood firm. McCarthy threatened. I braved him to do his
worst. We were to meet at the pool midway between our houses to talk it over.

"When I went down there I found him talking with his son, so I smoked a cigar and waited behind a
tree until he should be alone. But as I listened to his talk all that was black and bitter in me
seemed to come uppermost. He was urging his son to marry my daughter with as little regard for what
she might think as if she were a slut from off the streets. It drove me mad to think that I and all
that I held most dear should be in the power of such a man as this. Could I not snap the bond? I was
already a dying and a desperate man. Though clear of mind and fairly strong of limb, I knew that my
own fate was sealed. But my memory and my girl! Both could be saved if I could but silence that foul
tongue. I did it, Mr. Holmes. I would do it again. Deeply as I have sinned, I have led a life of
martyrdom to atone for it. But that my girl should be entangled in the same meshes which held me was
more than I could suffer. I struck him down with no more compunction than if he had been some foul
and venomous beast. His cry brought back his son; but I had gained the cover of the wood, though I
was forced to go back to fetch the cloak which I had dropped in my flight. That is the true story,
gentlemen, of all that occurred."

"Well, it is not for me to judge you," said Holmes as the old man signed the statement which had
been drawn out. "I pray that we may never be exposed to such a temptation."

"I pray not, sir. And what do you intend to do?"

"In view of your health, nothing. You are yourself aware that you will soon have to answer for your
deed at a higher court than the Assizes. I will keep your confession, and if McCarthy is condemned I
shall be forced to use it. If not, it shall never be seen by mortal eye; and your secret, whether
you be alive or dead, shall be safe with us."