Crime scene report #688847
"What have you to say now?"

Sherlock Holmes looked deeply chagrined. He drew a sovereign from his pocket and threw it down upon
the slab, turning away with the air of a man whose disgust is too deep for words. A few yards off he
stopped under a lamp-post and laughed in the hearty, noiseless fashion which was peculiar to him.