Crime scene report #642109
"Do you know, Watson," said he, "that it is one of the curses of a mind with a turn like mine that I
must look at everything with reference to my own special subject. You look at these scattered
houses, and you are impressed by their beauty. I look at them, and the only thought which comes to
me is a feeling of their isolation and of the impunity with which crime may be committed there."

"Good heavens!" I cried. "Who would associate crime with these dear old homesteads?"