Crime scene report #423167
The Black Swan is an inn of repute in the High Street, at no distance from the station, and there we
found the young lady waiting for us. She had engaged a sitting-room, and our lunch awaited us upon
the table.

"I am so delighted that you have come," she said earnestly. "It is so very kind of you both; but
indeed I do not know what I should do. Your advice will be altogether invaluable to me."