Crime scene report #622456
"Pray take a seat," said Holmes. "This is my friend and colleague, Dr. Watson, who is occasionally
good enough to help me in my cases. Whom have I the honour to address?"

"You may address me as the Count Von Kramm, a Bohemian nobleman. I understand that this gentleman,
your friend, is a man of honour and discretion, whom I may trust with a matter of the most extreme
importance. If not, I should much prefer to communicate with you alone."